requirements.txt
torch>=2.0,<3.0
facenet_pytorch>=2.5,<3.0
opencv-python>=4.8,<5.0
psycopg2-binary>=2.9,<3.0
faiss-cpu>=1.7,<2.0
prometheus_client>=0.17,<1.0
Pillow>=10.0,<11.0
numpy>=1.24,<2.0
cryptography>=42.0,<44.0
flask>=3.0,<4.0
